Chane Auth Middleware, Change Migrations

+++ app/Http/Controllers/TeamController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Team;
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Tymon\JWTAuth\Facades\JWTAuth;
 
 class TeamController extends Controller
 {
@@ -19,25 +21,39 @@ class TeamController extends Controller
     }
 
 
-    public function showAllTeams()
+    public function showAllTeams(Auth $auth)
     {
-
-        return response()->json(Team::all());
+        $teams = Team::with('users', 'owner')->whereHas('users', function ($u) {
+            return $u->where('user_id', \Auth::id());
+        })->orWhere('owner', \Auth::id())->get();
+        return response()->json($teams);
     }
 
     public function show($id)
     {
+        $team = Team::with('users', 'owner')
+            ->where('id', $id)
+            ->whereHas('users', function ($u) {
+                return $u->where('user_id', \Auth::id());
+            })
+            ->orWhere('owner', \Auth::id())
+            ->first();
 
-        return response()->json(Team::findOrFail($id));
+
+        return response()->json($team);
     }
 
     public function create(Request $request)
     {
+
         $this->validate($request, [
             'title' => 'required',
         ]);
 
-        $team = Team::create($request->all());
+        $team = new Team();
+        $team->title = $request->title;
+        $team->owner = \Auth::id();
+        $team->save();
 
         return response()->json($team, 201);
 
@@ -46,7 +62,10 @@ class TeamController extends Controller
 
     public function update($id, Request $request)
     {
-        $team = Team::findOrFail($id);
+        $team = Team::with('users')->where('owner',\Auth::id())->where('id',$id)->first();
+        if(!$team){
+            return response()->json(['failed','You have not access to edit this team'],401);
+        }
         $team->update($request->all());
 
         return response()->json($team, 200);
@@ -54,6 +73,10 @@ class TeamController extends Controller
 
     public function delete($id)
     {
+        $team = Team::with('users')->where('owner',\Auth::id())->where('id',$id)->first();
+        if(!$team){
+            return response()->json(['failed','You have not access to  delete this team'],401);
+        }
         Team::findOrFail($id)->delete();
         return response('Deleted Successfully', 200);
     }
